Stuff
Fix issues with Cheat Menu when language uses a comma to seperate decimals.
Fix translation issues with emotions command
Minimum version is ML 0.7.1
Made the Save Fixer apply automatically and made the Mod Menu preference disable it instead
enableCheatMenuButton is automatically on and can't be disabled
Enabled the weather command
Removed useless dlls from csproj

+++ SR2EssentialsMod/SR2EUtils.cs
@@ -723,7 +723,7 @@ namespace SR2E
         public static bool TryParseFloat(this SR2ECommand cmd, string input, out float value, float min, bool inclusive, float max)
         {
             value = 0;
-            try { value = int.Parse(input); }
+            try { value = float.Parse(input); }
             catch { return cmd.SendNotValidFloat(input); }
             if (inclusive)
             {
